Dunitz/experimental (#153)
* starting

* use logger, small refactor

* return errors

* drop me?

* update function signature

* lint

* importing

* add logging levels

* lint

* PR and fixes

* tweaks from code review
